Add styled navigation bar. Introduce an AppearanceHelper to share between extension and app
